Moved the footer to the app component.
There was a small delay from when the app gets placed in the dom - so
there was a little flash of content where the footer shows up, followed
by the app. So to just keep it simple, everything goes in appTemplate.

But this was a good lesson. Relying on WebWorkers for the initial state does mean I will have a little
delay. Obvious, but I needed to see it to "get it". A better solution
would be to expose the initial state to the UI and draw that first.